refactor(GameConfiguration): 숫자 3 -> static 전역 상수화(DIGITS_FOR_THIS_GAME)
숫자 3을 static 전역 상수화 하여, GameConfiguration 클래스에서 상수로 관리하고자 변경하였습니다.
혹여나 나중에 숫자를 4개로 늘리는 등의 변경 사항이 있을 수 있어, 변경에 유연하게 대처하고자 이와 같은 판단을 하였습니다.
추가적으로, 값 객체를 사용하여 좀 더 유연하게 만들것인가를 고려하고 있습니다.
