+++ src/main/java/baseball/config/GameConfiguration.java
@@ -1,0 +1,5 @@
+package baseball.config;
+
+public class GameConfiguration {
+    public static final int DIGITS_FOR_THIS_GAME = 3;
+}
